[#4] refactor: update as module
