+++ client/src/index.ts
@@ -1,11 +1,4 @@
 import App from './App'
-import DOMRenderer from './utils/renderDom'
+import { domRenderer } from './utils/wooact'
 
-console.log('loded')
-try {
-  const $target = document.getElementById('App')
-  console.log($target)
-  DOMRenderer.render(new App(), $target)
-} catch (e) {
-  console.error(e)
-}
+domRenderer(new App(), document.querySelector('#App'))
