+++ client/src/utils/wooact/domRenderer.ts
@@ -1,0 +1,7 @@
+import { Component } from './'
+
+const domRenderer = (element: Component<any, any>, container: HTMLElement) => {
+  container.appendChild(element.getElement())
+}
+
+export default domRenderer
