+++ client/src/utils/wooact/index.ts
@@ -1,0 +1,5 @@
+// import { Component } from './Component'
+// import './defaultElements'
+export { default as domRenderer } from './domRenderer'
+export { default as Component } from './Component'
+export {} from './defaultElements' // ?
